bluetooth/requestDevice/canonicalizeFilter/wrong-service-in-services-member.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="/bluetooth/resources/bluetooth-helpers.js"></script>
<script>
bluetooth_test(() => {
  let test_promises = Promise.resolve();
  generateRequestDeviceArgsWithServices(['wrong_service']).forEach(args => {
    test_promises = test_promises.then(() =>
       assert_promise_rejects_with_message(requestDeviceWithTrustedClick(args),
       new DOMException('Failed to execute \'requestDevice\' on \'Bluetooth\'' +
         ': Invalid Service name: \'wrong_service\'. It must be a valid UUID ' +
         'alias (e.g. 0x1234), UUID (lowercase hex characters e.g. ' +
         '\'00001234-0000-1000-8000-00805f9b34fb\'), or recognized standard ' +
         'name from https://www.bluetooth.com/specifications/gatt/services ' +
         'e.g. \'alert_notification\'.',
         new TypeError()),
       'Invalid service must reject the promise.'))
  });
  return test_promises;
}, 'Invalid service must reject the promise.');
</script>
